// src/api/__generated__/wallet/model/followListItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { FollowListItemResponseRelationStatus } from './followListItemResponseRelationStatus';

/**
 * 关注/粉丝/互关列表项
 */
export interface FollowListItemResponse {
  /** 用户ID */
  userId?: string;
  /** 用户昵称 */
  nickname?: string;
  /** 用户简介 */
  profile?: string;
  /** 用户头像URL */
  avatarUrl?: string;
  /** 用户状态：1-正常 0-禁用 */
  status?: string;
  /** 当前登录用户对该项用户的关系状态（未登录为 NONE） */
  relationStatus?: FollowListItemResponseRelationStatus;
}